fix(internal): Use unsigned int literals instead of casts
Also added a comment explaining why unsigned ints are used.

Issue: #6

+++ src/internal/sp_boolbuf_set.c
@@ -26,5 +26,6 @@ void sp_boolbuf_set(size_t idx, int val, void *buf)
 	if (val)
 		*byte |= ((unsigned char)1) << offset;
 	else
-		*byte &= (unsigned char)(~((unsigned int)1) << offset);
+		/* Negation works on int/uint or larger types (smaller ones get promoted) */
+		*byte &= (unsigned char)(~1U << offset);
 }
